+++ src/main/java/net/silentchaos512/loginar/block/urn/UrnData.java
@@ -13,11 +13,15 @@ public class UrnData {
     private final UrnTypes urnType;
     private final int clayColor;
     private final int gemColor;
-    private final NonNullList<ItemStack> items;
+    private NonNullList<ItemStack> items;
+    private NonNullList<ItemStack> upgrades;
 
     public static final String NBT_ROOT = "BlockEntityTag";
     public static final String NBT_CLAY_COLOR = "ClayColor";
     public static final String NBT_GEM_COLOR = "GemColor";
+    public static final String NBT_ITEMS = "Items";
+    public static final String NBT_UPGRADES = "Upgrades";
+
     public static final int DEFAULT_CLAY_COLOR = 0x985F45;
     public static final int DEFAULT_GEM_COLOR = 0x33EBCB;
 
@@ -26,6 +30,7 @@ public class UrnData {
         this.clayColor = clayColor;
         this.gemColor = gemColor;
         this.items = NonNullList.withSize(urnType.inventorySize(), ItemStack.EMPTY);
+        this.upgrades = NonNullList.withSize(urnType.upgradeSlots(), ItemStack.EMPTY);
     }
 
     public static UrnData fromItem(ItemStack stack) {
@@ -45,8 +50,11 @@ public class UrnData {
                 blockData.getInt(NBT_CLAY_COLOR),
                 blockData.getInt(NBT_GEM_COLOR)
         );
-        if (blockData.contains("Items", Tag.TAG_LIST)) {
+        if (blockData.contains(NBT_ITEMS, Tag.TAG_LIST)) {
             ContainerHelper.loadAllItems(blockData, ret.items);
+        }
+        if (blockData.contains(NBT_UPGRADES, Tag.TAG_LIST)) {
+            ContainerHelper.loadAllItems(blockData, ret.upgrades);
         }
         return ret;
     }
@@ -65,6 +73,12 @@ public class UrnData {
         tags.putInt(NBT_GEM_COLOR, gemColor);
     }
 
+    public void writeNbtToItem(ItemStack stack) {
+        CompoundTag tags = UrnHelper.getData(stack);
+        UrnHelper.saveAllItems(tags, NBT_ITEMS, items, false);
+        UrnHelper.saveAllItems(tags, NBT_UPGRADES, upgrades, false);
+    }
+
     public UrnTypes urnType() {
         return urnType;
     }
@@ -79,5 +93,22 @@ public class UrnData {
 
     public NonNullList<ItemStack> items() {
         return items;
+    }
+
+    public void setItems(NonNullList<ItemStack> list) {
+        this.items = list;
+    }
+
+    public NonNullList<ItemStack> upgrades() {
+        return upgrades;
+    }
+
+    public void addUpgrade(ItemStack stack) {
+        for (int i = 0; i < upgrades.size(); ++i) {
+            if (upgrades.get(i).isEmpty()) {
+                upgrades.set(i, stack);
+                return;
+            }
+        }
     }
 }