+++ src/main/java/net/silentchaos512/loginar/block/urn/UrnHelper.java
@@ -1,8 +1,15 @@
 package net.silentchaos512.loginar.block.urn;
 
+import net.minecraft.core.NonNullList;
 import net.minecraft.nbt.CompoundTag;
+import net.minecraft.nbt.ListTag;
+import net.minecraft.nbt.Tag;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.ItemLike;
+import net.silentchaos512.lib.util.NameUtils;
 import net.silentchaos512.loginar.setup.LsTags;
+import net.silentchaos512.loginar.setup.UrnTypes;
 
 public final class UrnHelper {
     private UrnHelper() {
@@ -11,6 +18,43 @@ public final class UrnHelper {
 
     public static boolean isUrn(ItemStack stack) {
         return stack.is(LsTags.Items.URNS);
+    }
+
+    public static boolean isUpgrade(ItemStack stack) {
+        return stack.is(LsTags.Items.URN_UPGRADES);
+    }
+
+    public static boolean hasUpgrade(ItemStack urn, ItemLike upgrade) {
+        // Efficiently check for a matching item ID
+        ResourceLocation upgradeId = NameUtils.fromItem(upgrade);
+        CompoundTag tags = getData(urn);
+        ListTag listTag = tags.getList("Upgrades", Tag.TAG_COMPOUND);
+
+        for (int i = 0; i < listTag.size(); ++i) {
+            CompoundTag compoundTag = listTag.getCompound(i);
+            ResourceLocation id = ResourceLocation.tryParse(compoundTag.getString("id"));
+            if (id != null && id.equals(upgradeId)) {
+                return true;
+            }
+        }
+
+        return false;
+    }
+
+    public static int getUpgradeCount(ItemStack urn) {
+        CompoundTag tags = getData(urn);
+        if (!tags.contains("Upgrades")) {
+            return 0;
+        }
+        return tags.getList("Upgrades", Tag.TAG_COMPOUND).size();
+    }
+
+    public static int getMaxUpgradeCount(ItemStack urn) {
+        UrnTypes type = UrnTypes.fromItem(urn);
+        if (type != null) {
+            return type.upgradeSlots();
+        }
+        return 0;
     }
 
     public static int getClayColor(ItemStack stack) {
@@ -31,7 +75,42 @@ public final class UrnHelper {
         getData(stack).putInt(UrnData.NBT_GEM_COLOR, color);
     }
 
-    private static CompoundTag getData(ItemStack stack) {
+    public static CompoundTag getData(ItemStack stack) {
         return stack.getOrCreateTagElement(UrnData.NBT_ROOT);
+    }
+
+    public static void loadAllItems(CompoundTag tag, String tagKey, NonNullList<ItemStack> items) {
+        // Taken from ContainerHelper, but can specify the list name
+        ListTag listtag = tag.getList(tagKey, 10);
+
+        for(int i = 0; i < listtag.size(); ++i) {
+            CompoundTag compoundtag = listtag.getCompound(i);
+            int j = compoundtag.getByte("Slot") & 255;
+            if (j >= 0 && j < items.size()) {
+                items.set(j, ItemStack.of(compoundtag));
+            }
+        }
+
+    }
+
+    public static CompoundTag saveAllItems(CompoundTag tag, String tagKey, NonNullList<ItemStack> items, boolean p_18979_) {
+        // Taken from ContainerHelper, but can specify the list name
+        ListTag listtag = new ListTag();
+
+        for(int i = 0; i < items.size(); ++i) {
+            ItemStack itemstack = items.get(i);
+            if (!itemstack.isEmpty()) {
+                CompoundTag compoundtag = new CompoundTag();
+                compoundtag.putByte("Slot", (byte)i);
+                itemstack.save(compoundtag);
+                listtag.add(compoundtag);
+            }
+        }
+
+        if (!listtag.isEmpty() || p_18979_) {
+            tag.put(tagKey, listtag);
+        }
+
+        return tag;
     }
 }