+++ src/main/java/net/silentchaos512/loginar/item/UpgradeItem.java
@@ -2,10 +2,12 @@ package net.silentchaos512.loginar.item;
 
 import net.minecraft.ChatFormatting;
 import net.minecraft.network.chat.Component;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.TooltipFlag;
 import net.minecraft.world.level.Level;
+import net.silentchaos512.lib.util.NameUtils;
 import net.silentchaos512.loginar.util.TextUtil;
 import org.jetbrains.annotations.Nullable;
 
@@ -23,6 +25,9 @@ public class UpgradeItem extends Item {
     public void appendHoverText(ItemStack stack, @Nullable Level level, List<Component> tooltip, TooltipFlag flags) {
         if (unimplemented) {
             tooltip.add(TextUtil.misc("not_implemented").withStyle(ChatFormatting.DARK_RED, ChatFormatting.ITALIC));
+        } else {
+            ResourceLocation name = NameUtils.fromItem(this);
+            tooltip.add(TextUtil.translate("item", name.getPath() + ".desc"));
         }
     }
 }